refactor: все папки src вернул в пакет ru.yandex.practicum.taskmanager
