+++ src/ru/yandex/practicum/taskmanager/tasks/Selftask.java
@@ -1,27 +1,28 @@
-package tasks;
+package ru.yandex.practicum.taskmanager.tasks;
 
-public class Epictask extends Task {
-    public Epictask(String name, String description) {
+public class Selftask extends Task {
+
+    public Selftask(String name, String description) {
         super(name, description);
     }
 
     @Override
     public Subordination getSubordination() {
-        return Subordination.EPIC;
-
+        return Subordination.SELF;
     }
 
     @Override
     public String toString() {
-        return "<Epic: id=" + super.getId() + "; name=" + super.getName() + "; desc=" +
+        return "<Standard: id=" + super.getId() + "; name=" + super.getName() + "; desc=" +
                 super.getDescription() + "; status=" + super.getStatus() + ">";
     }
 
     @Override
     public Task copy() {
-        Epictask copy = new Epictask(getName(), getDescription());
+        Selftask copy = new Selftask(getName(), getDescription());
         copy.setId(getId());
         copy.setStatus(getStatus());
         return copy;
     }
+
 }
