+++ src/ru/yandex/practicum/taskmanager/tasks/Subordination.java
@@ -1,4 +1,4 @@
-package tasks;
+package ru.yandex.practicum.taskmanager.tasks;
 
 public enum Subordination {
     SELF,
