+++ src/ru/yandex/practicum/taskmanager/utils/Generator.java
@@ -1,4 +1,4 @@
-package utils;
+package ru.yandex.practicum.taskmanager.utils;
 
 public interface Generator {
     Integer getId();
