+++ src/ru/yandex/practicum/taskmanager/utils/SerialGenerator.java
@@ -1,4 +1,4 @@
-package utils;
+package ru.yandex.practicum.taskmanager.utils;
 
 public class SerialGenerator implements Generator {
     private static int count;
